requirements.txt
pyrogram
tgcrypto
aiohttp
requests
pytz
python-dotenv
Flask>=3.0.0
gunicorn==20.1.0
pycryptodome
beautifulsoup4
cloudscraper
